1.6.2
* Update tests to be easliy maintainable (#43)

* Adds the ability to delete a worktree (#35)

* feat: initial functionality to delete a worktree

This adds functionality where the user can use a key binding (ctrl-d in
this instance) when the worktree is selected, and then delete the
worktree

* docs: add options to readme

* doc: split up config options

* fix: propergate keybinding correctly

* test: disable on dev branch for the moment

---------

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>

* feat: only select a commit if it's a new branch (#38)

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>

* feat: allow choosing of a branch or commit ref (#39)

You will now be given a list of branches on the third step, but given an
option to view commit refs instead

Changes:
 - Rename some variables to split out thier context
 - Set the initial list on the third step to be branches to select from

Adds:
 - `--bind` to the fzf call to allow the switch to view commit refs
instead of the branch
 - Documentation of the new config flags

closes: #36

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>

* mod: update keybinding usage

Set the keybinding like the delete keybinding for consistency

* build: remove test for now

This needs a rework and will be done as a separate ticket.

Due to the complexity of the testing script, it's hard to understand how
the tests work.

* mod: extract functionality into own files

* phase one

Extract functionality into own functions and contextual files

* phase one bug fix

* phase two

Add mocks and journey tests

* Phase three

Update tests

Notice:
Not all tests pass at this point

* tests: skip failing test

* test: correct tests

Changes:
 - Break up tests
 - Update tests where they were hanging

* mod: re-enable github workflow tests

* mod: re-enable tests on github .. again

---------

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>

* Update tests to be easliy maintainable (#44)

* Adds the ability to delete a worktree (#35)

* feat: initial functionality to delete a worktree

This adds functionality where the user can use a key binding (ctrl-d in
this instance) when the worktree is selected, and then delete the
worktree

* docs: add options to readme

* doc: split up config options

* fix: propergate keybinding correctly

* test: disable on dev branch for the moment

---------

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>

* feat: only select a commit if it's a new branch (#38)

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>

* feat: allow choosing of a branch or commit ref (#39)

You will now be given a list of branches on the third step, but given an
option to view commit refs instead

Changes:
 - Rename some variables to split out thier context
 - Set the initial list on the third step to be branches to select from

Adds:
 - `--bind` to the fzf call to allow the switch to view commit refs
instead of the branch
 - Documentation of the new config flags

closes: #36

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>

* mod: update keybinding usage

Set the keybinding like the delete keybinding for consistency

* build: remove test for now

This needs a rework and will be done as a separate ticket.

Due to the complexity of the testing script, it's hard to understand how
the tests work.

* mod: extract functionality into own files

* phase one

Extract functionality into own functions and contextual files

* phase one bug fix

* phase two

Add mocks and journey tests

* Phase three

Update tests

Notice:
Not all tests pass at this point

* tests: skip failing test

* test: correct tests

Changes:
 - Break up tests
 - Update tests where they were hanging

* mod: re-enable github workflow tests

* fix: issue with hanging test

---------

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>

* fix: remove chars from branch name (#46)

Changes:
 - Remove `+ ` from branch name when passing to the worktree add command

Adds:
 - Test around bug

closes: #41

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>

* Fix/47  issue with branches with forward slash in the branch name causing error (#48)

* fix: only clean up specific branch prefixes

Changes:
 - Remove `origin` and `remotes` prefixes from branches to align with
other clean up tasks

closes: #47

* mod: tidy up repitition of code

---------

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>

---------

Co-authored-by: Nigel Greenway <nigel.greenway@futurepixels.co.uk>